Move manageSession to utils
